# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrainingLog API",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/training/api"
    }
  ],
  "security": [
    {
      "ownerCookie": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ownerCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "TrainingLog.Auth"
      },
      "watchBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Scoped device token. Cookie authentication is not accepted on watch/v1."
      }
    },
    "schemas": {
...
      "WatchLinkStart": {
        "type": "object",
        "required": [
          "id",
          "token",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string",
            "pattern": "^[A-Fa-f0-9]{64}$",
            "description": "Client-generated 256-bit secret, saved in Keychain before the request. Never put in a URL."
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        }
      },
      "WatchLinkPoll": {
        "type": "object",
        "required": [
          "id",
          "token"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string",
            "pattern": "^[A-Fa-f0-9]{64}$"
          }
        }
      },
      "WatchLinkState": {
        "type": "object",
        "required": [
          "requestId",
          "label",
          "expiresAt",
          "state",
          "protocolVersion",
          "contractVersion"
        ],
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "expiresAt": {
            "type": "integer",
            "format": "int64"
          },
          "state": {
            "type": "string",
            "enum": [
              "pending",
              "approved"
            ]
          },
          "protocolVersion": {
            "type": "integer",
            "const": 1
          },
          "contractVersion": {
            "type": "integer",
            "const": 2
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/watch/link/window": {
      "post": {
        "summary": "Open a 120-second owner-authorized discovery window. Requires cookie and CSRF. Replaces unapproved requests.",
        "responses": {
          "200": {
            "description": "Window opened",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "expiresAt"
                  ],
                  "properties": {
                    "expiresAt": {
                      "type": "integer",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "CSRF validation failed"
          },
          "401": {
            "description": "Owner login required"
          },
          "429": {
            "description": "Rate limited"
          },
          "503": {
            "description": "Watch feature disabled"
          }
        },
        "operationId": "postWatchLinkWindow",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/watch/link/request": {
      "post": {
        "summary": "Request connection during an open window; identical id/secret retries retain the request. Does not grant access until phone approval.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchLinkStart"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchLinkState"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Previously approved device was revoked"
          },
          "409": {
            "description": "Window closed or request expired/reused"
          },
          "429": {
            "description": "Rate limited or three pending requests already exist"
          },
          "503": {
            "description": "Disabled"
          }
        },
        "operationId": "postWatchLinkRequest"
      }
    },
    "/watch/link/requests": {
      "get": {
        "summary": "List only this owner's unapproved requests. Compare the label shown on the physical watch before approving.",
        "responses": {
          "200": {
            "description": "Pending requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "label",
                      "expiresAt"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "label": {
                        "type": "string"
                      },
                      "expiresAt": {
                        "type": "integer",
                        "format": "int64"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Owner login required"
          }
        },
        "operationId": "getWatchLinkRequests"
      }
    },
    "/watch/link/requests/{id}/approve": {
      "post": {
        "summary": "Explicitly approve matching physical watch. Requires owner cookie/CSRF. Idempotent; closes discovery and discards other unapproved requests.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Approved; the previously generated secret now authenticates this device",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "approved"
                  ],
                  "properties": {
                    "approved": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "CSRF validation failed"
          },
          "401": {
            "description": "Owner login required"
          },
          "409": {
            "description": "Request expired, foreign or device limit reached"
          },
          "503": {
            "description": "Disabled"
          }
        },
        "operationId": "postWatchLinkRequestsIdApprove"
      }
    },
    "/watch/link/status": {
      "post": {
        "summary": "Poll with the persisted secret every two seconds. Approved requests remain recoverable after a lost response; revoked devices cannot reconnect this way.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchLinkPoll"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request state; expiresAt becomes the 90-day token expiry upon approval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchLinkState"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Revoked device"
          },
          "410": {
            "description": "Unknown, expired or incorrect-secret request"
          },
          "429": {
            "description": "Poll rate limited"
          }
        },
        "operationId": "postWatchLinkStatus"
      }
    },
...
  }
}